+++ source/renderer/app/stores/HardwareWalletsStore.js
@@ -58,6 +58,7 @@ import {
   prepareTrezorOutput,
   prepareTrezorCertificate,
   prepareTrezorWithdrawal,
+  prepareTrezorAuxiliaryData,
 } from '../utils/shelleyTrezor';
 import {
   DeviceModels,
@@ -1701,6 +1702,15 @@ export default class HardwareWalletsStore extends Store {
       prepareTrezorCertificate(certificate)
     );
 
+    let auxiliaryData = null;
+    if (this.votingData) {
+      const { votingKey, nonce } = this.votingData;
+      auxiliaryData = prepareTrezorAuxiliaryData({
+        votingKey,
+        nonce: nonce.toString(),
+      });
+    }
+
     const recognizedDevice = find(
       this.hardwareWalletDevices,
       (hardwareWalletDevice) => hardwareWalletDevice.paired === walletId
@@ -1764,6 +1774,7 @@ export default class HardwareWalletsStore extends Store {
         certificates: certificatesData,
         withdrawals: withdrawalsData,
         devicePath: recognizedDevicePath,
+        auxiliaryData,
       });
 
       if (!signedTransaction.success) {